benchmarks/requirements.txt
grpcio==1.78.0
grpcio-reflection==1.78.0
protobuf==6.33.6
pydantic==2.12.5
fastapi==0.135.1
uvicorn==0.41.0
httpx>=0.28.0
psycopg2-binary>=2.9.9
